Solitaire board class was splitted in two classes.
