+++ src/eu/veldsoft/four/row/solitaire/CardStack.java
@@ -1,7 +1,7 @@
 /*
  This file is a part of Four Row Solitaire
 
- Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov
+ Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov, Vanya Gyaurova, Plamena Popova, Hristiana Kalcheva
 
  Four Row Solitaire is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -25,6 +25,18 @@ import java.awt.Point;
 import java.util.Vector;
 
 import javax.swing.JLayeredPane;
+
+/**
+ * 
+ * @author Todor Balabanov
+ */
+class CardLayeredPane extends JLayeredPane {
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+}
 
 /**
  * Class: CardStack
