+++ src/eu/veldsoft/four/row/solitaire/GameActivity.java
@@ -1,7 +1,7 @@
 /*
  This file is a part of Four Row Solitaire
 
- Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov
+ Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov, Vanya Gyaurova, Plamena Popova, Hristiana Kalcheva
 
  Four Row Solitaire is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -19,7 +19,11 @@
 
 package eu.veldsoft.four.row.solitaire;
 
+import javax.swing.JOptionPane;
+
 import android.app.Activity;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
@@ -37,11 +41,11 @@ import android.widget.Toast;
  */
 public class GameActivity extends Activity {
 	// TODO Find better object model.
-	
+
 	/**
 	 * 
 	 */
-	private SolitaireBoard board = null;
+	private SolitaireBoard board = null;//new SolitaireBoard();
 
 	/**
 	 * On creation.
@@ -68,15 +72,48 @@ public class GameActivity extends Activity {
 	public boolean onOptionsItemSelected(MenuItem item) {
 		switch (item.getItemId()) {
 		case R.id.new_game:
-			board.newGame(GameState.GAME_LOST);
+			/*
+			 * Ask the user for new game confirmation.
+			 */
+			// AlertDialog.Builder builder = new AlertDialog.Builder(this);
+			//
+			// builder.setTitle("Confirm");
+			// builder.setMessage("Are you sure?");
+			//
+			// builder.setPositiveButton("YES",
+			// new DialogInterface.OnClickListener() {
+			//
+			// public void onClick(DialogInterface dialog, int which) {
+			// dialog.dismiss();
+			// }
+			//
+			// });
+			//
+			// builder.setNegativeButton("NO",
+			// new DialogInterface.OnClickListener() {
+			//
+			// @Override
+			// public void onClick(DialogInterface dialog, int which) {
+			// // Do nothing
+			// dialog.dismiss();
+			// }
+			// });
+			//
+			// AlertDialog alert = builder.create();
+			// alert.show();
+
+			if (true) {
+				board.recordGame(GameState.GAME_LOST, 0, 0, 0, 0, false);
+				board.newGame(GameState.GAME_LOST);
+			}
 			break;
 		case R.id.undo_last_move:
-			Toast.makeText(GameActivity.this, "Undo", Toast.LENGTH_SHORT)
-					.show();
+			board.undoMove();
 			break;
 		case R.id.hint:
-			Toast.makeText(GameActivity.this, "Hint", Toast.LENGTH_SHORT)
-					.show();
+			String hint[] = board.getHint();
+			Toast.makeText(GameActivity.this, hint[0] + " " + hint[1],
+					Toast.LENGTH_SHORT).show();
 			break;
 		}
 		return true;
