+++ src/eu/veldsoft/four/row/solitaire/SolitaireBoard.java
@@ -1,7 +1,7 @@
 /*
  This file is a part of Four Row Solitaire
 
- Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov
+ Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov, Vanya Gyaurova, Plamena Popova, Hristiana Kalcheva
 
  Four Row Solitaire is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -19,82 +19,12 @@
 
 package eu.veldsoft.four.row.solitaire;
 
-//import java.awt.BorderLayout;
-//import java.awt.FlowLayout;
-//import java.awt.event.ActionEvent;
-//import java.awt.event.ActionListener;
-//import java.awt.event.MouseEvent;
-//import java.awt.event.WindowAdapter;
-//import java.awt.event.WindowEvent;
-import java.awt.HeadlessException;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.util.LinkedList;
-
-//import javax.swing.ImageIcon;
-import javax.swing.JFrame;
-//import javax.swing.JLabel;
-//import javax.swing.JOptionPane;
-//import javax.swing.JPanel;
-
-//import javax.swing.Timer;
-//import javax.swing.event.MouseInputAdapter;
-
-/**
- * 
- * @author Konstantin Tsanov
- * 
- */
-// TODO Get it out as separate Java file.
-class SolitaireFrame extends JFrame {
-	
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-	/**
-	 * Solitaire Board.
-	 */
-	private SolitaireBoard board = new SolitaireBoard();
-
-	/**
-	 * Timer.
-	 */
-	private Timer timer = new Timer(1000, new TimerListener());
-
-	/**
-	 * Status bar.
-	 */
-	private JPanel statusBar = new JPanel(new FlowLayout(FlowLayout.CENTER));
-
-	/**
-	 * Timer label.
-	 */
-	private JLabel timerLabel = new JLabel("Time: OFF");
-
-	/**
-	 * 
-	 * @throws HeadlessException
-	 */
-	public SolitaireFrame() throws HeadlessException {
-		super();
-		setTitle("Four Row Solitaire");
-		setSize(806, 700);
-		setLocationRelativeTo(null);
-		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-		setResizable(false);
-		setIconImage(new ImageIcon(getClass().getResource("images/logo.png"))
-				.getImage());
-
-		setVisible(true);
-
-		addWindowListener(wl);
-	}
-}
 
 /**
  * Class: SolitaireBoard
@@ -111,19 +41,9 @@ class SolitaireBoard {
 	private static final int INITIAL_CARDS_NUMBER_IN_COLUMN = 5;
 
 	/**
-	 * Find better OOP modeling alternative! Use enumerated type for card back.
-	 */
-	static int deckNumber = 3;
-
-	/**
 	 * Can be 1 or 3. Should be only here!
 	 */
 	static int drawCount = 1;
-
-	/**
-	 * Find better OOP modeling alternative! Use enumerated type for card back.
-	 */
-	static int backgroundNumber = 2;
 
 	/**
 	 * To store new option selection for next new game, otherwise the count
@@ -134,57 +54,32 @@ class SolitaireBoard {
 	/**
 	 * The four columns for the main playing field.
 	 */
-	private Column[] columns = new Column[4];
+	// TODO Should be private.
+	Column[] columns = new Column[4];
 
 	/**
 	 * The discard pile.
 	 */
-	private DiscardPile discardPile = new DiscardPile();
+	// TODO Should be private.
+	DiscardPile discardPile = new DiscardPile();
 
 	/**
 	 * The deal pile.
 	 */
-	private DealDeck dealDeck = new DealDeck(discardPile);
+	// TODO Should be private.
+	DealDeck dealDeck = new DealDeck(discardPile);
 
 	/**
 	 * The four ace piles (to stack Ace - King of a single suit).
 	 */
-	private AcePile[] acePiles = new AcePile[4];
+	// TODO Should be private.
+	AcePile[] acePiles = new AcePile[4];
 
 	/**
 	 * The four top individual cells.
 	 */
-	private SingleCell[] cells = new SingleCell[4];
-
-	/**
-	 * 
-	 */
-	private SolitairePanel mainPanel;
-
-	/**
-	 * 
-	 */
-	private MyMouseListener ml = new MyMouseListener();
-
-	/**
-	 * 
-	 */
-	public MyWindowListener wl = new MyWindowListener();
-
-	/**
-	 * Timer count.
-	 */
-	private int timerCount = 0;
-
-	/**
-	 * Timer to next game.
-	 */
-	private int timerToRunNextGame = 0;
-
-	/**
-	 * Timer to run.
-	 */
-	private boolean timerToRun = false;
+	// TODO Should be private.
+	SingleCell[] cells = new SingleCell[4];
 
 	/**
 	 * 1 = easy, 2 = medium, 3 = hard Should be only here!
@@ -199,22 +94,25 @@ class SolitaireBoard {
 	/**
 	 * Source.
 	 */
-	private LinkedList<CardStack> sourceList = new LinkedList<CardStack>();
+	// TODO Should be private.
+	LinkedList<CardStack> sourceList = new LinkedList<CardStack>();
 
 	/**
 	 * Destination.
 	 */
-	private LinkedList<CardStack> destinationList = new LinkedList<CardStack>();
+	// TODO Should be private.
+	LinkedList<CardStack> destinationList = new LinkedList<CardStack>();
 
 	/**
 	 * Card numbers.
 	 */
-	private LinkedList<Integer> numCards = new LinkedList<Integer>();
+	// TODO Should be private.
+	LinkedList<Integer> numCards = new LinkedList<Integer>();
 
 	/**
 	 * The cards from the discard pile.
 	 */
-	private LinkedList<Integer> numCardsInDiscardView = new LinkedList<Integer>();
+	LinkedList<Integer> numCardsInDiscardView = new LinkedList<Integer>();
 
 	/**
 	 * Sets the board's window name, size, location, close button option, makes
@@ -227,40 +125,18 @@ class SolitaireBoard {
 	 * Creates the solitaire board.
 	 * 
 	 * @param cards
-	 * 
 	 * @param numViewableCards
+	 * 
+	 * @author Todor Balabanov 
 	 */
 	public void createBoard(LinkedList<Integer> cards, int numViewableCards) {
-		mainPanel = new SolitairePanel();
-		mainPanel.setLayout(new SolitaireLayout());
-
-		mainPanel.changeBackground(backgroundNumber);
-
 		for (int i = 0; i < columns.length; i++) {
 			columns[i] = new Column();
-			columns[i].addMouseListener(ml);
 		}
-
-		mainPanel.add(columns[0], SolitaireLayout.COLUMN_ONE);
-		mainPanel.add(columns[1], SolitaireLayout.COLUMN_TWO);
-		mainPanel.add(columns[2], SolitaireLayout.COLUMN_THREE);
-		mainPanel.add(columns[3], SolitaireLayout.COLUMN_FOUR);
 
 		for (int i = 0; i < cells.length; i++) {
 			cells[i] = new SingleCell();
-			cells[i].addMouseListener(ml);
 		}
-
-		mainPanel.add(cells[0], SolitaireLayout.CELL_ONE);
-		mainPanel.add(cells[1], SolitaireLayout.CELL_TWO);
-		mainPanel.add(cells[2], SolitaireLayout.CELL_THREE);
-		mainPanel.add(cells[3], SolitaireLayout.CELL_FOUR);
-
-		dealDeck.addMouseListener(ml);
-		discardPile.addMouseListener(ml);
-
-		mainPanel.add(dealDeck, SolitaireLayout.DECK);
-		mainPanel.add(discardPile, SolitaireLayout.DISCARD_PILE);
 
 		for (int i = 0; i < acePiles.length; i++) {
 			switch (i) {
@@ -280,21 +156,7 @@ class SolitaireBoard {
 			default:
 				break;
 			}
-
-			acePiles[i].addMouseListener(ml);
 		}
-
-		mainPanel.add(acePiles[0], SolitaireLayout.SPADES_ACE_PILE);
-		mainPanel.add(acePiles[1], SolitaireLayout.CLUBS_ACE_PILE);
-		mainPanel.add(acePiles[2], SolitaireLayout.DIAMONDS_ACE_PILE);
-		mainPanel.add(acePiles[3], SolitaireLayout.HEARTS_ACE_PILE);
-
-		statusBar.add(timerLabel);
-
-		JPanel p1 = new JPanel(new BorderLayout());
-		p1.add(mainPanel, BorderLayout.CENTER);
-		p1.add(statusBar, BorderLayout.SOUTH);
-		add(p1);
 
 		if (cards != null) {
 			dealOutCustomBoard(cards, numViewableCards);
@@ -305,8 +167,10 @@ class SolitaireBoard {
 
 	/**
 	 * Deals the cards.
+	 * 
+	 * @author Todor Balabanov
 	 */
-	private void dealOutBoard() {
+	void dealOutBoard() {
 		LinkedList<Card> cards = (LinkedList<Card>) Deck.getFullShuffledDeck();
 
 		/*
@@ -341,21 +205,6 @@ class SolitaireBoard {
 		}
 
 		dealDeck.setDeck(cards);
-
-		timerCount = 0;
-
-		if (timerToRunNextGame == 1) {
-			timer.stop();
-			timerLabel.setText("Timer: 0");
-			timerToRun = true;
-		} else {
-			timer.stop();
-			timerLabel.setText("Time: OFF");
-
-			timerToRun = false;
-		}
-
-		mainPanel.revalidate();
 	}
 
 	/**
@@ -364,8 +213,10 @@ class SolitaireBoard {
 	 * @param numbers
 	 * 
 	 * @param numViewableCards
+	 * 
+	 * @author Todor Balabanov
 	 */
-	private void dealOutCustomBoard(LinkedList<Integer> numbers,
+	void dealOutCustomBoard(LinkedList<Integer> numbers,
 			int numViewableCards) {
 		LinkedList<Card> cards = (LinkedList<Card>) Deck
 				.getDeckSubsetByCardNumbers(numbers);
@@ -400,20 +251,6 @@ class SolitaireBoard {
 		}
 
 		discardPile.setView(numViewableCards);
-
-		if (timerToRunNextGame == 1) {
-			timer.stop();
-			timerLabel.setText("Time: " + (timerCount == -1 ? 0 : timerCount));
-
-			timerToRun = true;
-		} else {
-			timer.stop();
-			timerLabel.setText("Time: OFF");
-
-			timerToRun = false;
-		}
-
-		mainPanel.revalidate();
 	}
 
 	/**
@@ -464,26 +301,6 @@ class SolitaireBoard {
 	 */
 	public void newGame(GameState winOrLoss) {
 		/*
-		 * If the game was won, the win was already reported.
-		 */
-		if (winOrLoss != GameState.GAME_WON
-				&& winOrLoss != GameState.DO_NOTHING) {
-			int check = JOptionPane.showConfirmDialog(this,
-					"Quitting the current game will result in a loss.\n"
-							+ "Do you wish to continue?", "Continue?",
-					JOptionPane.PLAIN_MESSAGE);
-
-			if (check == JOptionPane.YES_OPTION) {
-				recordGame(GameState.GAME_LOST);
-			} else {
-				/*
-				 * If player wants to continue game.
-				 */
-				return;
-			}
-		}
-
-		/*
 		 * Remove cards from ace piles. Set numTimesThroughDeck back to 1.
 		 */
 		clearBoard();
@@ -497,25 +314,18 @@ class SolitaireBoard {
 	}
 
 	/**
-	 * Used to reset the stats.
-	 */
-	public void resetStats() {
-		recordGame(GameState.RESET_STATS);
-	}
-
-	/**
-	 * Save options.
-	 */
-	public void saveOptions() {
-		recordGame(GameState.DO_NOTHING);
-	}
-
-	/**
 	 * Manages the game states.
 	 * 
 	 * @param winOrLoss
+	 * @param timerCount 
+	 * @param backgroundNumber 
+	 * @param deckNumber 
+	 * @param timerToRunNextGame
+	 * @param timerToRun
+	 * 
+	 * @author Todor Balabanov
 	 */
-	private void recordGame(GameState winOrLoss) {
+	void recordGame(GameState winOrLoss, int deckNumber, int backgroundNumber, int timerCount, int timerToRunNextGame, boolean timerToRun) {
 		int count = 0, temp = 0;
 		int gamesPlayed1e = 0, gamesWon1e = 0, winStreak1e = 0, lossStreak1e = 0, currentStreak1e = 0;
 		int gamesPlayed1m = 0, gamesWon1m = 0, winStreak1m = 0, lossStreak1m = 0, currentStreak1m = 0;
@@ -1040,20 +850,6 @@ class SolitaireBoard {
 	}
 
 	/**
-	 * Manages the appaearance.
-	 * 
-	 * @param deck
-	 * 
-	 * @param background
-	 */
-	public void setAppearance(int deck, int background) {
-		deckNumber = deck;
-		backgroundNumber = background;
-
-		mainPanel.changeBackground(backgroundNumber);
-	}
-
-	/**
 	 * Returns the draw count.
 	 * 
 	 * @return
@@ -1095,99 +891,6 @@ class SolitaireBoard {
 		if (newDrawCount != 3 && newDrawCount != 1) {
 			newDrawCount = 1;
 		}
-	}
-
-	/**
-	 * Returns the deck number.
-	 * 
-	 * @return
-	 */
-	public int getDeckNumber() {
-		return deckNumber;
-	}
-
-	/**
-	 * Sets the deck number.
-	 * 
-	 * @param deckNum
-	 */
-	public void setDeckNumber(int deckNum) {
-		deckNumber = deckNum;
-
-		if (deckNumber > ChangeAppearance.NUM_DECKS || deckNumber <= 0) {
-			deckNumber = ChangeAppearance.FRS_DECK;
-		}
-	}
-
-	/**
-	 * Returns the background number.
-	 * 
-	 * @return
-	 */
-	public int getBackgroundNumber() {
-		return backgroundNumber;
-	}
-
-	/**
-	 * Sets the background number.
-	 * 
-	 * @param backNum
-	 */
-	public void setBackgroundNumber(int backNum) {
-		backgroundNumber = backNum;
-
-		if (backgroundNumber > ChangeAppearance.NUM_BACKGROUNDS
-				|| backgroundNumber <= 0) {
-			backgroundNumber = ChangeAppearance.FRS_BACKGROUND;
-		}
-	}
-
-	/**
-	 * Returns next game timer status.
-	 * 
-	 * @return
-	 */
-	public int getTimerNextGameStatus() {
-		return timerToRunNextGame;
-	}
-
-	/**
-	 * Returns the timer status.
-	 * 
-	 * @return
-	 */
-	public int getTimerStatus() {
-		if (timer.isRunning()) {
-			return 1;
-		} else {
-			return 0;
-		}
-	}
-
-	/**
-	 * Sets timer status.
-	 * 
-	 * @param timerInt
-	 */
-	public void setTimerStatus(int timerInt) {
-		if (timerInt == 1) {
-			timerToRunNextGame = 1;
-		} else if (timerInt == 0) {
-			timerToRunNextGame = 0;
-		}
-
-		if (timerInt != 0 && timerInt != 1) {
-			timerToRunNextGame = 0;
-		}
-	}
-
-	/**
-	 * Sets the timer counter.
-	 * 
-	 * @param time
-	 */
-	public void setTimer(int time) {
-		timerCount = time;
 	}
 
 	/**
@@ -1242,6 +945,7 @@ class SolitaireBoard {
 		if (sourceList.isEmpty()) {
 			return;
 		}
+		
 		/*
 		 * If player is holding on to a card.
 		 */
@@ -1258,20 +962,11 @@ class SolitaireBoard {
 			if (num == 1) {
 				discardPile.setView(numDiscard);
 				tempSource.peek().unhighlight();
-
-				ml.clickedCard = null;
-				ml.hasSelected = false;
-				ml.singleCardSelected = false;
-				ml.temp = null;
 			} else {
 				for (int i = 0; i < num; i++) {
 					tempSource.getCardAtLocation(tempSource.length() - i - 1)
 							.unhighlight();
 				}
-
-				ml.clickedCard = null;
-				ml.hasSelected = false;
-				ml.temp = null;
 			}
 
 			tempSource.repaint();
@@ -1343,9 +1038,13 @@ class SolitaireBoard {
 
 	/**
 	 * Manages the hints.
+	 * 
+	 * @return Hint structure.
+	 * 
+	 * @author Todor Balabanov
 	 */
 	@SuppressWarnings("fallthrough")
-	public void getHint() {
+	public String[] getHint() {
 		CardStack source = new CardStack();
 		CardStack destination = new CardStack();
 		CardStack temp = new CardStack();
@@ -1529,472 +1228,18 @@ class SolitaireBoard {
 			}
 		}
 
+		String hint[] = { "", "" };
 		if (hints.isEmpty() == false) {
-			String string = "";
-
 			for (int i = 0; i < hints.size(); i++) {
-				string += hints.get(i) + "\n";
+				hint[0] += hints.get(i) + "\n";
 			}
-
-			JOptionPane.showMessageDialog(this, string, "Hints Galore",
-					JOptionPane.INFORMATION_MESSAGE);
+			hint[1] = "Hints Galore";
 		} else {
-			JOptionPane.showMessageDialog(this,
-					"There are no moves on the field.\n"
-							+ "Either deal more cards or start a new game",
-					"Hints", JOptionPane.INFORMATION_MESSAGE);
-		}
-	}
-
-	/**
-	 * Manages the mouse events.
-	 * 
-	 * @author Todor Balabanov
-	 * 
-	 */
-	private class MyMouseListener extends MouseInputAdapter {
-		
-		/**
-		 * If true, the player hasn't completed a move.
-		 */
-		private boolean hasSelected = false;
-
-		/**
-		 * If true, the selected stack is only one card.
-		 */
-		private boolean singleCardSelected = false;
-
-		/**
-		 * 
-		 */
-		private Card clickedCard;
-		
-		/**
-		 * 
-		 */
-		private CardStack source;
-		
-		/**
-		 * 
-		 */
-		private CardStack destination;
-		
-		/**
-		 * 
-		 */
-		private CardStack temp;
-
-		/**
-		 * For right clicking discard pile view.
-		 */
-		private Card tempCard;
-
-		/**
-		 * To prevent clicking cards from the right click view.
-		 */
-		private boolean rightClicked = false;
-
-		/**
-		 * Checks if the game is won
-		 */
-		private void checkWin() {
-			for (int i = 0; i < acePiles.length; i++) {
-				if (acePiles[i].isEmpty()
-						|| acePiles[i].peek().getNumber().equals(CardRank.KING) == false) {
-					return;
-				}
-			}
-
-			if (WinScreen.animation != 0 || WinScreen.sounds != 0) {
-				new WinScreen();
-			}
-
-			if (timerToRun) {
-				final TopTimes top = new TopTimes();
-				int pos = top.IsTopTime(timerCount);
-
-				top.addWindowListener(new WindowAdapter() {
-					@Override
-					public void windowClosing(WindowEvent e) {
-						top.setVisible(false);
-						int playAgain = JOptionPane.showConfirmDialog(
-								SolitaireBoard.this, "Play Again?", "You Won!",
-								JOptionPane.YES_NO_OPTION);
-
-						if (playAgain == JOptionPane.YES_OPTION) {
-							recordGame(GameState.GAME_WON);
-							newGame(GameState.GAME_WON);
-						} else if (playAgain == JOptionPane.NO_OPTION) {
-							recordGame(GameState.GAME_WON);
-							System.exit(0);
-						}
-					}
-				});
-
-				if (pos >= 0) {
-					top.setProperties(timerCount);
-					top.setVisible(true);
-				} else {
-					top.dispose();
-				}
-			}
-
-			int playAgain = JOptionPane.showConfirmDialog(SolitaireBoard.this,
-					"Play Again?", "You Won!", JOptionPane.YES_NO_OPTION);
-
-			if (playAgain == JOptionPane.YES_OPTION) {
-				recordGame(GameState.GAME_WON);
-				newGame(GameState.GAME_WON);
-			} else if (playAgain == JOptionPane.NO_OPTION) {
-				recordGame(GameState.GAME_WON);
-				System.exit(0);
-			}
+			hint[0] = "There are no moves on the field.\n"
+					+ "Either deal more cards or start a new game";
+			hint[1] = "Hints";
 		}
 
-		/**
-		 * Mouse-pressed event.
-		 * 
-		 * @param e
-		 */
-		public void mousePressed(MouseEvent e) {
-			if (e.getButton() == MouseEvent.BUTTON3
-					&& e.getSource() == discardPile) {
-				if (discardPile.getNumViewableCards() == 1
-						|| (discardPile.getNumViewableCards() == 0 && !discardPile
-								.isEmpty())) {
-					tempCard = discardPile.pop();
-					discardPile.repaint();
-					rightClicked = true;
-				}
-			}
-		}
-
-		/**
-		 * Mouse-released event.
-		 * 
-		 * @param e
-		 */
-		public void mouseReleased(MouseEvent e) {
-			if (e.getButton() == MouseEvent.BUTTON3 && tempCard != null) {
-				discardPile.push(tempCard);
-				discardPile.repaint();
-				rightClicked = false;
-				tempCard = null;
-			}
-		}
-
-		/**
-		 * Mouse-clicked event.
-		 * 
-		 * @param e
-		 */
-		public void mouseClicked(MouseEvent e) {
-			discardPile.repaint();
-			discardPile.revalidate();
-
-			if (!timer.isRunning() && timerToRun) {
-				timer.start();
-			}
-
-			if ((e.getButton() != MouseEvent.BUTTON1) || rightClicked) {
-				return;
-			}
-
-			else if (e.getClickCount() == 2 && hasSelected
-					&& singleCardSelected) {
-				if (source.peek().getNumber().equals(CardRank.ACE)) {
-					if (source.peek().getSuit().equals(CardSuit.SPADES)) {
-						Card card = source.pop();
-						card.unhighlight();
-						acePiles[0].push(card);
-						destinationList.add(acePiles[0]);
-					} else if (source.peek().getSuit().equals(CardSuit.CLUBS)) {
-						Card card = source.pop();
-						card.unhighlight();
-						acePiles[1].push(card);
-						destinationList.add(acePiles[1]);
-					} else if (source.peek().getSuit()
-							.equals(CardSuit.DIAMONDS)) {
-						Card card = source.pop();
-						card.unhighlight();
-						acePiles[2].push(card);
-						destinationList.add(acePiles[2]);
-					} else if (source.peek().getSuit().equals(CardSuit.HEARTS)) {
-						Card card = source.pop();
-						card.unhighlight();
-						acePiles[3].push(card);
-						destinationList.add(acePiles[3]);
-					}
-
-					hasSelected = false;
-					source.repaint();
-					repaint();
-					return;
-				}
-
-				for (int i = 0; i < acePiles.length; i++) {
-					if (!acePiles[i].isEmpty()
-							&& source.peek().getSuit()
-									.equals(acePiles[i].peek().getSuit())
-							&& source
-									.peek()
-									.getNumber()
-									.isLessByOneThan(
-											(acePiles[i].peek().getNumber()))) {
-						Card card = source.pop();
-						card.unhighlight();
-						acePiles[i].push(card);
-
-						destinationList.add(acePiles[i]);
-						hasSelected = false;
-
-						source.repaint();
-						repaint();
-
-						if (card.getNumber().equals(CardRank.KING)) {
-							checkWin();
-						}
-
-						return;
-					}
-				}
-
-				for (int i = 0; i < cells.length; i++) {
-					if (cells[i].isEmpty()) {
-						Card card = source.pop();
-						card.unhighlight();
-						cells[i].push(card);
-
-						destinationList.add(cells[i]);
-						hasSelected = false;
-
-						source.repaint();
-						repaint();
-						return;
-					}
-				}
-
-				source.peek().unhighlight();
-
-				source.repaint();
-				repaint();
-				return;
-			} else if (e.getClickCount() == 2 && hasSelected) {
-				hasSelected = false;
-
-				if (temp.length() > 0) {
-					for (int i = 0; i < temp.length(); i++) {
-						source.getCardAtLocation(source.length() - i - 1)
-								.unhighlight();
-					}
-				}
-
-				sourceList.removeLast();
-				numCardsInDiscardView.removeLast();
-				numCards.removeLast();
-			}
-
-			else if (!hasSelected && e.getClickCount() == 1
-					|| (e.getSource() instanceof DealDeck)) {
-				source = (CardStack) e.getSource();
-
-				if (source instanceof DealDeck) {
-					if (hasSelected) {
-						hasSelected = false;
-
-						if (temp.length() > 0) {
-							for (int i = 0; i < temp.length(); i++) {
-								sourceList
-										.getLast()
-										.getCardAtLocation(
-												sourceList.getLast().length()
-														- i - 1).unhighlight();
-							}
-						} else {
-							sourceList.getLast().peek().unhighlight();
-						}
-
-						sourceList.getLast().repaint();
-						repaint();
-						sourceList.removeLast();
-						numCardsInDiscardView.removeLast();
-						numCards.removeLast();
-					}
-
-					numCardsInDiscardView
-							.add(discardPile.getNumViewableCards());
-					clickedCard = source.pop();
-
-					if (clickedCard != null) {
-						sourceList.add(dealDeck);
-						destinationList.add(discardPile);
-						numCards.add(discardPile.getNumViewableCards());
-					}
-					/*
-					 * The deck was reset but the player hasn't used up the
-					 * times through the deck.
-					 */
-					else if (dealDeck.hasDealsLeft()) {
-						sourceList.add(dealDeck);
-						destinationList.add(discardPile);
-						numCards.add(0);
-					} else {
-						numCardsInDiscardView.removeLast();
-					}
-
-					return;
-				}
-
-				numCardsInDiscardView.add(discardPile.getNumViewableCards());
-				clickedCard = source.getCardAtLocation(e.getPoint());
-
-				if (clickedCard != null) {
-					hasSelected = true;
-					temp = source.getStack(clickedCard);
-
-					sourceList.add(source);
-					numCards.add(temp.length());
-
-					if (temp.length() > 1) {
-						singleCardSelected = false;
-					} else {
-						singleCardSelected = true;
-					}
-				} else {
-					numCardsInDiscardView.removeLast();
-					hasSelected = false;
-					return;
-				}
-			}
-			/*
-			 * Stack/card already selected.
-			 */
-			else if (e.getClickCount() == 1 && hasSelected) {
-				destination = (CardStack) e.getSource();
-
-				if (singleCardSelected) {
-					if (destination.isValidMove(clickedCard)) {
-						Card card = source.pop();
-						card.unhighlight();
-						destination.push(card);
-
-						/*
-						 * If move is valid, add destination info for undo.
-						 */
-						destinationList.add(destination);
-
-						if (destination instanceof AcePile
-								&& clickedCard.getNumber()
-										.equals(CardRank.KING)) {
-							repaint();
-							checkWin();
-						}
-					} else {
-						/*
-						 * Not needed with highlighting version.
-						 */
-						source.peek().unhighlight();
-
-						/*
-						 * Upon invalid move, remove undo information for cards.
-						 */
-						sourceList.removeLast();
-						numCards.removeLast();
-						numCardsInDiscardView.removeLast();
-					}
-				} else {
-					if (destination.isValidMove(temp)) {
-						CardStack stack = new CardStack();
-
-						for (int i = temp.length(); i > 0; i--) {
-							Card card = source.pop();
-							card.unhighlight();
-
-							stack.push(card);
-						}
-
-						destination.push(stack);
-
-						/*
-						 * If move is valid, add destination info for undo.
-						 */
-						destinationList.add(destination);
-					} else {
-						for (int i = temp.length() - 1; i >= 0; i--) {
-							source.getCardAtLocation(source.length() - i - 1)
-									.unhighlight();
-						}
-
-						/*
-						 * Upon invalid move, remove undo information for cards.
-						 */
-						sourceList.removeLast();
-						numCards.removeLast();
-						numCardsInDiscardView.removeLast();
-					}
-				}
-
-				singleCardSelected = false;
-				hasSelected = false;
-				temp = null;
-				clickedCard = null;
-			}
-
-			repaint();
-		}
-	}
-
-	/**
-	 * Timer displaying.
-	 * 
-	 * @author Todor Balabanov
-	 * 
-	 */
-	private class TimerListener implements ActionListener {
-
-		/**
-		 * Action performed
-		 * 
-		 * @param e
-		 */
-		public void actionPerformed(ActionEvent e) {
-			if (e.getSource() != timer) {
-				return;
-			}
-			timerCount++;
-			timerLabel.setText("Time: " + timerCount);
-			statusBar.repaint();
-		}
-	}
-
-	/**
-	 * Manages the window events.
-	 * 
-	 * @author Todor Balabanov
-	 * 
-	 */
-	public class MyWindowListener extends WindowAdapter {
-
-		/**
-		 * On closing the main window:
-		 * 
-		 * @param e
-		 */
-		public void windowClosing(WindowEvent e) {
-			int save = JOptionPane
-					.showConfirmDialog(
-							SolitaireBoard.this,
-							"Closing without saving will result "
-									+ "in a loss, would you like to save the current game?",
-							"Save Game?", JOptionPane.YES_NO_OPTION);
-
-			if (save == JOptionPane.YES_OPTION) {
-				recordGame(GameState.GAME_SAVED);
-				System.exit(0);
-			} else if (save == JOptionPane.NO_OPTION) {
-				recordGame(GameState.GAME_LOST);
-				System.exit(0);
-			}
-		}
+		return (hint);
 	}
 }
