+++ src/eu/veldsoft/four/row/solitaire/SolitairePanel.java
@@ -1,7 +1,7 @@
 /*
  This file is a part of Four Row Solitaire
 
- Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov
+ Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov, Vanya Gyaurova, Plamena Popova, Hristiana Kalcheva
 
  Four Row Solitaire is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -51,7 +51,7 @@ class SolitairePanel extends JPanel {
 	public SolitairePanel() {
 		URL imageURL = this.getClass().getResource(
 				"images/backgrounds/background"
-						+ SolitaireBoard.backgroundNumber + ".jpg");
+						+ SolitaireFrame.backgroundNumber + ".jpg");
 
 		if (imageURL != null) {
 			background = new ImageIcon(imageURL).getImage();
@@ -65,7 +65,7 @@ class SolitairePanel extends JPanel {
 	 * @param back
 	 */
 	public void changeBackground(int back) {
-		SolitaireBoard.backgroundNumber = back;
+		SolitaireFrame.backgroundNumber = back;
 
 		URL imageURL = this.getClass().getResource(
 				"images/backgrounds/background" + back + ".jpg");
