+++ src/eu/veldsoft/four/row/solitaire/StatisticsActivity.java
@@ -1,7 +1,7 @@
 /*
  This file is a part of Four Row Solitaire
 
- Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov
+ Copyright (C) 2010-2014 by Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov, Vanya Gyaurova, Plamena Popova, Hristiana Kalcheva
 
  Four Row Solitaire is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -35,7 +35,7 @@ import android.widget.TabHost.TabSpec;
  * @author Matt Stephen, Todor Balabanov, Konstantin Tsanov, Ventsislav Medarov
  * 
  */
-public class StatisticsActivity extends Activity {
+public class StatisticsActivity extends TabActivity {
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
