+++ src/eu/veldsoft/four/row/solitaire/StatisticsTab3Activity.java
@@ -9,6 +9,10 @@ import android.view.View.OnClickListener;
 import android.widget.Button;
 import android.widget.Toast;
 
+/**
+ * 
+ * @author Ina Baltadzhieva
+ */
 public class StatisticsTab3Activity extends Activity {
 
 	@Override
