+++ src/eu/veldsoft/four/row/solitaire/StatisticsTab4Activity.java
@@ -9,6 +9,10 @@ import android.view.View.OnClickListener;
 import android.widget.Button;
 import android.widget.Toast;
 
+/**
+ * 
+ * @author Ina Baltadzhieva
+ */
 public class StatisticsTab4Activity extends Activity {
 
 	@Override
